refactor: migrated some Stellar functions from horizon file to logic file
