+++ src/families/stellar/api/index.js
@@ -1,3 +1,3 @@
 //@flow
 
-export { getAccount, getOperations } from "./horizon";
+export { getAccount, getOperations, fetchBaseFee } from "./horizon";
